requirements.txt
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.13.3
aiohttp==3.9.3
web3==6.15.1
eth-utils==3.0.0
eth-typing==4.0.0
